pyproject.toml
[project]
name = "flasklogin"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask-restx>=1.3.0",
    "flask-sqlalchemy>=3.1.1",
    "flask>=3.1.0",
    "mysql-connector-python>=9.2.0",
    "pymysql>=1.1.1",
    "python-dotenv>=1.0.1",
]
